requirements.txt
# =============================================================================
# Banking Lakehouse - Python Dependencies
# =============================================================================

# Core Spark & Delta Lake
pyspark==3.4.0
delta-spark==2.4.0

# Data Processing
pandas==2.0.3
pyarrow==12.0.1

# Data Quality
great-expectations==0.17.19

# dbt (SQL Transformations)
dbt-core==1.6.0
dbt-spark==1.6.0

# Data Generation
faker==19.3.0

# Cloud / S3 (MinIO)
boto3==1.28.25

# Orchestration
apache-airflow==2.7.0

# Kafka
confluent-kafka==2.2.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
chispa==0.9.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.0
